some cleaning and fixing range cardinalities for select

+++ hyracks-fullstack/hyracks/hyracks-storage-am-statistics/src/main/java/org/apache/hyracks/storage/am/statistics/sketch/quantile/QuantileSketch.java
@@ -207,7 +207,7 @@ public class QuantileSketch<T extends Comparable<T>> implements ISketch<T, T> {
         return elements.size();
     }
 
-    public int length() {
+    public double length() {
         return size;
     }
